Modified compose-test.sh to run gridmet container.

+++ compose-test.sh
@@ -13,6 +13,15 @@ set -e
 
 . ./nhm.env			# environment variables file
 
+# encapsulate some container running boiler-plate
+run () {
+    svc=$1
+
+    echo ""
+    echo "Running $svc..."
+    docker-compose $COMPOSE_FILES -p nhm run --rm $svc
+} # run
+
 echo "Building necessary Docker images"
 docker-compose build base_image
 docker-compose build --parallel
@@ -37,10 +46,18 @@ fi
 COMPOSE_FILES="-f docker-compose.yml -f docker-compose-testing.yml"
 echo "Beginning run. If this fails at any point, run the following command:"
 echo "docker-compose $COMPOSE_FILES down"
-for svc in data_loader ofp ncf2cbh nhm-prms out2ncf verifier nhm-restart; do
-    echo ""
-    echo "Running ${svc}..."
-    docker-compose $COMPOSE_FILES -p nhm run --rm $svc
+
+# call run() function above
+run data_loader
+run gridmet
+
+if [ $? != 0 ]; then
+    echo 'Gridmet data is not available - process exiting'
+    exit 1
+fi
+
+for svc in ofp ncf2cbh nhm-prms out2ncf verifier nhm-restart; do
+    run "$svc"
 done
 docker-compose $COMPOSE_FILES down
